Towards private contacts vs public contacts
