+++ app/src/assets/countries.js
@@ -1,4 +1,5 @@
 export const countryList = [
+  { name: "Non Affiliate", code: "N/A" },
   { name: "Afghanistan", code: "AF" },
   { name: "land Islands", code: "AX" },
   { name: "Albania", code: "AL" },
