+++ app/src/components/organizations/OrgUtils.js
@@ -13,7 +13,7 @@ const countryNames = countryList.map((c) => c.name);
 
 export const orgTypes = ["CNA", "ADP", "BULK_DOWNLOAD", "SECRETARIAT"];
 
-export function emptyContactInfo() {
+export function emptyPrivateContacts() {
   return {
     poc: "",
     poc_email: "",
