+++ app/src/components/organizations/OrganizationEditPage.vue
@@ -11,34 +11,35 @@
         <v-card-title class="cve-background"> Organization </v-card-title>
         <v-divider></v-divider>
         <v-card-text>
-          <template v-for="field in editableFields" :key="field.key">
-            <FieldInput
-              :field="field"
-              :model-value="localOrg[field.key]"
-              :pending-value="orgStore.currentOrg.review?.[field.key]"
-              :has-diff="hasDiff(field.key) && !dismissed.has(field.key)"
-              @update:model-value="localOrg[field.key] = $event"
-              :all-users="
-                field.type === 'user' || field.type === 'users'
-                  ? field.allUsers
-                  : undefined
-              "
-              @accept="acceptReview(field.key)"
-              @deny="denyReview(field.key)"
-            />
-          </template>
+          <v-row>
+            <v-col
+              v-for="field in editableFields"
+              :key="field.key"
+              cols="12"
+              :lg="field.lg ?? 12"
+            >
+              <FieldInput
+                :field="field"
+                :model-value="localOrg[field.key]"
+                :pending-value="orgStore.currentOrg.review?.[field.key]"
+                :has-diff="hasDiff(field.key) && !dismissed.has(field.key)"
+                @update:model-value="localOrg[field.key] = $event"
+                :all-users="
+                  field.type === 'user' || field.type === 'users'
+                    ? field.allUsers
+                    : undefined
+                "
+                @accept="acceptReview(field.key)"
+                @deny="denyReview(field.key)"
+              />
+            </v-col>
+          </v-row>
         </v-card-text>
       </v-card>
 
       <MembersView :editable="true"></MembersView>
 
-      <v-card class="mb-5">
-        <v-card-title class="cve-background">Contact Information</v-card-title>
-        <v-divider></v-divider>
-        <v-card-text>
-          <ContactInfoForm v-model="localOrg.contact_info" />
-        </v-card-text>
-      </v-card>
+      <PrivateContactsForm v-model="localOrg.contact_info" />
 
       <ProgramDataForm
         v-if="auth.isSecretariat"
@@ -68,17 +69,16 @@ import { ref, watch, computed, onMounted } from "vue";
 import { useRouter } from "vue-router";
 import { useOrgStore } from "../../stores/organizationStore.js";
 import FieldInput from "../inputs/FieldInput.vue";
-import OrganizationContactInfoInput from "./contact_info/ContactInfoForm.vue";
 import MembersView from "./MembersView.vue";
 import { useAuthStore } from "../../stores/loginAuth.js";
 import ProgramDataForm from "./program_data/ProgramDataForm.vue";
-import ContactInfoForm from "./contact_info/ContactInfoForm.vue";
+import PrivateContactsForm from "./contact_info/PrivateContactsForm.vue";
 
 import {
   getFieldsForOrg,
   normalizeFieldDef,
   ROLES,
-  emptyContactInfo,
+  emptyPrivateContacts,
 } from "./OrgUtils.js";
 
 const auth = useAuthStore();
@@ -103,7 +103,10 @@ const EXCLUDED_KEYS = new Set([
 function normalizeOrg(o = {}) {
   const copy = { ...o };
 
-  copy.contact_info = { ...emptyContactInfo(), ...(copy.contact_info || {}) };
+  copy.contact_info = {
+    ...emptyPrivateContacts(),
+    ...(copy.contact_info || {}),
+  };
   copy.program_data = { ...(copy.program_data || {}) };
 
   if (!Array.isArray(copy.admins)) copy.admins = [];
