+++ app/src/components/organizations/OrganizationViewPage.vue
@@ -195,7 +195,7 @@
     </v-row>
 
     <!-- Contact Info -->
-    <ContactInfoView
+    <PrivateContactsView
       :contact_info="org.contact_info"
       :review="reviewOrNull?.contact_info"
     />
@@ -239,7 +239,7 @@ import { useRouter } from "vue-router";
 import { useOrgStore } from "../../stores/organizationStore.js";
 import { useAuthStore } from "../../stores/loginAuth.js";
 import AuditTrail from "./audits/AuditTrail.vue";
-import ContactInfoView from "./contact_info/ContactInfo.vue";
+import PrivateContactsView from "./contact_info/PrivateContacts.vue";
 import Conversations from "./conversations/ConversationView.vue";
 import MembersView from "./MembersView.vue";
 import { getFieldsForOrg, ROLES } from "./OrgUtils.js";
