+++ app/src/components/organizations/OrganizationsCreatePage.vue
@@ -20,16 +20,60 @@
               />
             </v-col>
           </v-row>
+
+          <v-col cols="12" sm="6">
+            <v-text-field
+              v-model="model.poc"
+              label="Point of contact"
+              variant="outlined"
+              density="comfortable"
+              hide-details="auto"
+            />
+          </v-col>
+          <v-col cols="12" sm="6">
+            <v-text-field
+              v-model="model.poc_email"
+              label="POC email"
+              type="email"
+              variant="outlined"
+              density="comfortable"
+              hide-details="auto"
+              :rules="[rules.email]"
+            />
+          </v-col>
+          <v-col cols="12" sm="6">
+            <v-text-field
+              v-model="model.phone"
+              label="Phone"
+              type="tel"
+              variant="outlined"
+              density="comfortable"
+              hide-details="auto"
+              :rules="[rules.phone]"
+            />
+          </v-col>
+          <v-col cols="12" sm="6">
+            <v-text-field
+              v-model="model.website"
+              label="Website"
+              type="url"
+              variant="outlined"
+              density="comfortable"
+              hide-details="auto"
+              :rules="[rules.uri]"
+            />
+          </v-col>
         </v-card-text>
       </v-card>
 
-      <v-card class="">
-        <v-card-title class="cve-background">Contact Information</v-card-title>
-        <v-divider></v-divider>
-        <v-card-text>
-          <ContactInfoForm v-model="organization.contact_info" />
-        </v-card-text>
-      </v-card>
+      <PrivateContactsForm v-model="organization.contact_info" />
+
+      <ProgramDataForm
+        v-if="auth.isSecretariat"
+        :program_data="organization.program_data"
+        @update:program_data="organization.program_data = $event"
+      />
+
       <v-card-actions class="px-4 pb-4">
         <v-spacer />
         <v-btn variant="text" @click="resetForm">Reset</v-btn>
@@ -40,25 +84,33 @@
 </template>
 
 <script setup>
+//Standard imports
 import { ref, computed } from "vue";
 import { useRouter } from "vue-router";
+const router = useRouter();
+
+//Stores
 import { useOrgStore } from "../../stores/organizationStore.js";
 import { useAuthStore } from "../../stores/loginAuth.js";
+import { useGlobalDialogStore } from "../../stores/globalDialogStore.js";
+const dialogStore = useGlobalDialogStore();
+const orgStore = useOrgStore();
+const auth = useAuthStore();
+
+//Components
 import FieldInput from "../inputs/FieldInput.vue";
-import ContactInfoForm from "./contact_info/ContactInfoForm.vue";
+import PrivateContactsForm from "./contact_info/PrivateContactsForm.vue";
 import GICSCodeInput from "../inputs/GICSCodeInput.vue";
+import ProgramDataForm from "./program_data/ProgramDataForm.vue";
+
+//Composables
 import {
   getFieldsForOrg,
   normalizeFieldDef,
   ROLES,
-  emptyContactInfo,
+  emptyPrivateContacts,
 } from "./OrgUtils.js";
-import { useGlobalDialogStore } from "../../stores/globalDialogStore.js";
 
-const dialogStore = useGlobalDialogStore();
-const router = useRouter();
-const orgStore = useOrgStore();
-const auth = useAuthStore();
 const form = ref(null);
 const isFormValid = computed(() => form.value?.isValid === true);
 
@@ -78,7 +130,8 @@ const defaultEmpty = () => ({
   hard_quota: null,
   authority: ["CNA"],
   industry: "",
-  contact_info: emptyContactInfo(),
+  contact_info: emptyPrivateContacts(),
+  program_data: {},
 });
 
 const organization = ref(defaultEmpty());
