--- app/src/components/organizations/contact_info/ContactInfoForm.vue
@@ -1,145 +1,0 @@
-<template>
-  <v-row v-if="model">
-    <!-- Main contact fields -->
-    <v-col cols="12" sm="6">
-      <v-text-field
-        v-model="model.poc"
-        label="Point of contact"
-        variant="outlined"
-        density="comfortable"
-        hide-details="auto"
-      />
-    </v-col>
-    <v-col cols="12" sm="6">
-      <v-text-field
-        v-model="model.poc_email"
-        label="POC email"
-        type="email"
-        variant="outlined"
-        density="comfortable"
-        hide-details="auto"
-        :rules="[rules.email]"
-      />
-    </v-col>
-    <v-col cols="12" sm="6">
-      <v-text-field
-        v-model="model.phone"
-        label="Phone"
-        type="tel"
-        variant="outlined"
-        density="comfortable"
-        hide-details="auto"
-        :rules="[rules.phone]"
-      />
-    </v-col>
-    <v-col cols="12" sm="6">
-      <v-text-field
-        v-model="model.website"
-        label="Website"
-        type="url"
-        variant="outlined"
-        density="comfortable"
-        hide-details="auto"
-        :rules="[rules.uri]"
-      />
-    </v-col>
-
-    <!-- Additional contacts -->
-    <v-col cols="12">
-      <div class="d-flex align-center justify-space-between mb-2">
-        <span class="text-subtitle-2">Additional contacts</span>
-        <v-btn size="small" variant="tonal" color="primary" @click="addContact">
-          + Add
-        </v-btn>
-      </div>
-
-      <v-row
-        v-for="(contact, index) in model.additional_contacts"
-        :key="index"
-        class="mb-1 align-center"
-      >
-        <v-col cols="12" sm="4">
-          <v-text-field
-            v-model="contact.poc"
-            label="Name"
-            variant="outlined"
-            density="comfortable"
-            hide-details="auto"
-          />
-        </v-col>
-        <v-col cols="12" sm="4">
-          <v-text-field
-            v-model="contact.poc_email"
-            label="Email"
-            type="email"
-            variant="outlined"
-            density="comfortable"
-            hide-details="auto"
-            :rules="[rules.email]"
-          />
-        </v-col>
-        <v-col cols="12" sm="3">
-          <v-text-field
-            v-model="contact.phone"
-            label="Phone"
-            type="tel"
-            variant="outlined"
-            density="comfortable"
-            hide-details="auto"
-            :rules="[rules.phone]"
-          />
-        </v-col>
-        <v-col cols="12" sm="1" class="d-flex justify-center">
-          <v-btn
-            icon="mdi-minus"
-            size="small"
-            variant="tonal"
-            color="error"
-            @click="removeContact(index)"
-          />
-        </v-col>
-      </v-row>
-    </v-col>
-  </v-row>
-</template>
-
-<script setup>
-import { computed } from "vue";
-import { getFieldRules } from "../../inputs/validators.js";
-
-const props = defineProps({
-  modelValue: { type: Object, required: true },
-});
-
-const emit = defineEmits(["update:modelValue"]);
-
-const model = computed({
-  get: () => props.modelValue,
-  set: (val) => emit("update:modelValue", val),
-});
-
-const rules = {
-  email: getFieldRules({ validator: { name: "email" } })[0] ?? true,
-  phone: getFieldRules({ validator: { name: "phone" } })[0] ?? true,
-  uri: getFieldRules({ validator: { name: "uri" } })[0] ?? true,
-};
-
-function addContact() {
-  emit("update:modelValue", {
-    ...props.modelValue,
-    additional_contacts: [
-      ...(props.modelValue.additional_contacts ?? []),
-      { poc: "", poc_email: "", phone: "" },
-    ],
-  });
-}
-
-function removeContact(index) {
-  const updated = [...props.modelValue.additional_contacts];
-  updated.splice(index, 1);
-  emit("update:modelValue", {
-    ...props.modelValue,
-    additional_contacts: updated,
-  });
-}
-</script>
