+++ app/src/components/organizations/program_data/ProgramDataForm.vue
@@ -1,7 +1,7 @@
 <template>
   <v-card class="mb-4">
     <v-card-title class="text-subtitle-1 cve-background"
-      >Program Info</v-card-title
+      ><span>Program Info (Secreteriat only)</span></v-card-title
     >
     <v-divider />
     <v-card-text>
